Move StructureSetItem subclasses to subfolder
