+++ proknow-sdk/Patient/Entities/StructureSet/StructureSetVersionItem.cs
@@ -8,7 +8,7 @@ using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 using ProKnow.Exceptions;
 
-namespace ProKnow.Patient.Entities
+namespace ProKnow.Patient.Entities.StructureSet
 {
     /// <summary>
     /// Represents a version of a structure set
