+++ proknow-sdk/Patient/Entities/StructureSetItem.cs
@@ -8,6 +8,7 @@ using System.Text.Json;
 using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 using ProKnow.Exceptions;
+using ProKnow.Patient.Entities.StructureSet;
 
 namespace ProKnow.Patient.Entities
 {
